requirements.txt
anyio==3.7.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.2
fastapi==0.99.1
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-speech==2.21.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
idna==3.4
numpy==1.25.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.27.0
typing-extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
